feat(vm): replace numeric values with scalar model

+++ crates/vela_reflect/src/members.rs
@@ -41,7 +41,7 @@ pub fn name(registry: &TypeRegistry, target: &ReflectValue) -> ReflectResult<Ref
 
 pub fn id(registry: &TypeRegistry, target: &ReflectValue) -> ReflectResult<ReflectValue> {
     match target_type(registry, target) {
-        Ok(desc) => Ok(ReflectValue::Host(HostValue::Int(
+        Ok(desc) => Ok(ReflectValue::Host(HostValue::i64(
             // TODO(reflect): stable IDs are u64, but reflection currently exposes IDs
             // through signed script ints. Replace this lossy saturation with a deliberate
             // unsigned/ID value surface before treating reflect::id() as a stable public
